% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Add}
\alias{Add}
\alias{Add.OptionPos}
\title{S3 generic function  Add a object input a list}
\usage{
Add(object, ...)

\method{Add}{OptionPos}(object, strike, expiration, type = c("C", "P"),
  underlyingName = "", name = "", multiplier = 100, amount = 0,
  price = NA, delta = NA, vega = NA, theta = NA, underlyingPrice = NA,
  underlyingSigma = NA, interestRate = NA, ...)
}
\arguments{
\item{object}{the object to be working on}

\item{...}{Other parameters}

\item{strike}{The strick of the new option}

\item{expiration}{The expiration date of the new option}

\item{type}{What is the type of the option: call or put}

\item{underlyingName}{The underlying ticker of the option}

\item{name}{The name of the option, if missing automatically construct one based on the option information}

\item{multiplier}{The contract size of the option}

\item{amount}{The quantity of the option contracts}

\item{price}{The current price of the option}

\item{delta}{The delta of the option}

\item{vega}{The vega of the option}

\item{theta}{The theta of the option}

\item{underlyingPrice}{The price of the underlying security}

\item{underlyingSigma}{The volatility of the underlying security}

\item{interestRate}{the current interestedRate}
}
\value{
The OptionPos with the new option appended
}
\description{
S3 generic function  Add a object input a list
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPos}: Append an new option object to the end of OptionPos object
}}

