% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AddMaxDrawDownArrow.R
\name{AddMaxDrawdownArrow}
\alias{AddMaxDrawdownArrow}
\title{Add maximum drawdown}
\usage{
AddMaxDrawdownArrow(perf, mdd)
}
\arguments{
\item{perf}{The performance data}

\item{mdd}{The maximum drawdown calcauted from the performance data}
}
\value{
Trivial values
}
\description{
Add the maximum draw down chart onto a given plot. (Does not work with plot.xts)
}
\note{
see \code{\link{MaxDrawdown}} for examples
}

