% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AlignXts.R
\name{AlignPnl}
\alias{AlignPerf}
\alias{AlignPnl}
\title{Align two Pnl time sereies}
\usage{
AlignPnl(pnl1, pnl2)

AlignPerf(perf1, perf2)
}
\arguments{
\item{pnl1}{A xts object}

\item{pnl2}{Another xts object}

\item{perf1}{A xts object}

\item{perf2}{Another xts object}
}
\value{
A table with two pnl aligned

A table with two performance aligned
}
\description{
Align two Pnl time sereies

Algin two performance time series
}

