% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DeltaAdjNotional}
\alias{DeltaAdjNotional}
\alias{DeltaAdjNotional.OptionPos}
\title{Delta adjust notional value of a portfolio}
\usage{
DeltaAdjNotional(object, ...)

\method{DeltaAdjNotional}{OptionPos}(object)
}
\arguments{
\item{object}{An OptionPos object}
}
\value{
the Delta Adjust Notional value
}
\description{
Delta adjust notional value of a portfolio

Delta Adjust notional value of an OptionPos object
}

