% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EquityPtf.R
\name{EquityPtf}
\alias{EquityPtf}
\title{EquityPtf Class (S3) Constructor: this is a portfolio of equities}
\usage{
EquityPtf()
}
\value{
Generate empty EquityPtf object
}
\description{
EquityPtf Class (S3) Constructor: this is a portfolio of equities
}

