% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetHistoricalMembers.R
\name{GetHistoricalMembers}
\alias{GetHistoricalMembers}
\title{Get the historical members from Bloomberg}
\usage{
GetHistoricalMembers(indexName, dates, firstn = 0)
}
\arguments{
\item{indexName}{The Bloomberg Ticker of the index Name}

\item{dates}{The given dates that the histrical members will be returned}

\item{firstn}{Return the first n numbers if n > 0}
}
\value{
A list of the members sorted by Percentage weight (if possible)
}
\description{
Get the historical members from Bloomberg
}
\examples{
\dontrun{
GetHistoricalMembers("SPX Index", Sys.Date())
}
}

