% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetMemberHistory.R
\name{GetMemberHistory}
\alias{GetMemberHistory}
\title{Import member history Index Data}
\usage{
GetMemberHistory(indexName, fields, stDate, endDate, monthly = F,
  firstn = 0)
}
\arguments{
\item{indexName}{The ticker of the index, e.g. SPX Index}

\item{fields}{A vector of fields}

\item{stDate}{Starting date}

\item{endDate}{Ending date}

\item{monthly}{If montly data is preferred}

\item{firstn}{If specified, only return the top fistn members with highest weight}
}
\value{
A list including the following items
\itemize{
\item symbols
\item fields
\item data
\item indexData
}
}
\description{
Import member history Index Data
}
\note{
This function requires a lot bloomberg data if the period between stDate
and endDate are very long
}

