% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSymbols.Bloomberg.R
\name{getSymbols.Bloomberg}
\alias{getSymbols.Bloomberg}
\title{Get symbols from Bloomberg. This function is not suppose to be called directly, but to be called by getSymbols of quantmod}
\usage{
getSymbols.Bloomberg(Symbols, env, return.class = "xts",
  from = as.Date("2007-01-01"), to = Sys.Date(), bb.suffix = "Equity",
  bb.interval = "5", ...)
}
\arguments{
\item{Symbols}{a list of Bloomberg tickers, for consistency with other load symbols method in quantmod, it should be only the name of the tickers}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{class of returned object}

\item{from}{Retrieve no earlier than this date}

\item{to}{Retrieve though this date}

\item{bb.suffix}{The Bloomberg yellow ticker}

\item{bb.interval}{Time interval for intraday data}

\item{...}{Other parameters passed in by getSymbols}
}
\value{
Depends on auto.assing
}
\description{
Get symbols from Bloomberg. This function is not suppose to be called directly, but to be called by getSymbols of quantmod
}

