% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggChartSeries}
\alias{ggChartSeries}
\title{ggplot version of Candle chart}
\usage{
ggChartSeries(x, start = NA, end = NA)
}
\arguments{
\item{x}{An OHLC object to be charted}

\item{start}{The starting time}

\item{end}{The ending time}
}
\description{
ggplot version of Candle chart
}

