% Generated by roxygen2 (4.0.1): do not edit by hand
\name{InformationRatio}
\alias{InformationRatio}
\title{Infromation ratio}
\usage{
InformationRatio(pnl, indexPnl, freq = 1, na.rm = F)
}
\arguments{
\item{pnl}{The P&L of the performance}

\item{indexPnl}{The P&L of the benchmark}

\item{freq}{The sample frequency}

\item{na.rm}{A boolean value to show if NA's should be removed during calcuation}
}
\value{
Infromation Ratio: \url{https://en.wikipedia.org/wiki/Information_ratio}
}
\description{
Infromation ratio
}

