% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.OptionPos}
\alias{is.OptionPos}
\title{Check if an object is OptionPos}
\usage{
is.OptionPos(object)
}
\arguments{
\item{object}{An object to be examed}
}
\value{
A boolean value to indicated if object is OptionPos
}
\description{
Check if an object is OptionPos
}

