% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.OptionPtf.R
\name{is}
\alias{is}
\alias{is.EquityPtf}
\alias{is.OptionPtf}
\title{Check if an object is Corresponding object}
\usage{
is.OptionPtf(object)

is.EquityPtf(object)
}
\arguments{
\item{object}{An object to be examed}
}
\value{
A boolean value to indicated if object is OptionPtf

A boolean value to indicated if object is OptionPtf
}
\description{
Check if an object is Corresponding object
}

