% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Notional}
\alias{Notional}
\alias{Notional.OptionPos}
\title{The Notional value of a portfolio}
\usage{
Notional(object, ...)

\method{Notional}{OptionPos}(object, ...)
}
\arguments{
\item{object}{The portfolio object}

\item{...}{Other parameters}
}
\value{
The notioal value
}
\description{
The Notional value of a portfolio

The notional value of the given option portfolio object
}

