% Generated by roxygen2 (4.0.1): do not edit by hand
\name{OptionPos}
\alias{OptionPos}
\title{OptionPos Class (S3) Constructor: this is a portfolio of options}
\usage{
OptionPos()
}
\value{
Generate empty OptionPos object
}
\description{
OptionPos Class (S3) Constructor: this is a portfolio of options
}

