% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OptionPtf.R
\name{OptionPtf}
\alias{OptionPtf}
\title{OptionPtf Class (S3) Constructor: this is a portfolio of options}
\usage{
OptionPtf()
}
\value{
Generate empty OptionPtf object
}
\description{
OptionPtf Class (S3) Constructor: this is a portfolio of options
}

