% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Performance2Pnl.R
\name{Performance2Pnl}
\alias{Performance2Pnl}
\title{P&L from performance}
\usage{
Performance2Pnl(perf, firstDayPNL = 0)
}
\arguments{
\item{perf}{Performance or the cumulative Peformance}

\item{firstDayPNL}{the P&L of first day}
}
\value{
P&L
}
\description{
P&L from performance
}

