% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerformanceSummary.R
\name{PerformanceSummary}
\alias{AnnualizedPerformanceSummary}
\alias{MonthlyPerformanceSummary}
\alias{PerformanceSummary}
\alias{YearlyPerformanceSummary}
\title{A summary of Performance}
\usage{
PerformanceSummary(pnl, indexPnl, r = 0)

MonthlyPerformanceSummary(dates, pnl, indexPnl, r = 0)

YearlyPerformanceSummary(dates, pnl, indexPnl, r = 0)

AnnualizedPerformanceSummary(pnl, indexPnl, r = 0, freq = 1/12)
}
\arguments{
\item{pnl}{The P&L}

\item{indexPnl}{The P&L of benchmark, used to calucated the information ratio}

\item{r}{Interest rate}

\item{dates}{The corresponding dates of the P&L}

\item{freq}{The sample frequency}
}
\value{
A summary of the performance, include the following names:
\itemize{
       \item Return The total return
       \item Sd The standard derivation
       \item Mdd The maximium draw down
       \item SharpeRatio The Sharpe ratio
       \item IR The information ratio
       \item PostiveRate The postive performance ratio
       \item tValue The student t value
       \item pValue The p value assuming two tail test. For one tail, divided by 2
       }
}
\description{
A summary of Performance

The month performance summary if the corresponding dates are given

The yearly performance summary if the corresponding dates are given

The Annulized performance summary if the sample frequency is given
}
\examples{
\dontrun{
yahoo <- getSymbols('YHOO',auto.assign = F)
perf <- Ad(yahoo)
pnl <- Performance2Pnl(perf)
spy <- getSymbols('SPY',auto.assign = F)
indexPnl <- Performance2Pnl(Ad(spy))
pnls <- AlignPnl(pnl, indexPnl)
PerformanceSummary(pnls[,1], pnls[,2], r= 0.003)
YearlyPerformanceSummary(index(pnls), pnls[,1], pnls[,2], r=0.003)
AnnualizedPerformanceSummary(pnls[,1], pnls[,2], r=0.003, freq=1/265)
}
}

