% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PerformanceTable.R
\name{PerformanceTable}
\alias{PerformanceTable}
\title{Summarize the performance of serveral portfolio}
\usage{
PerformanceTable(res, perf, size, index.pnl)
}
\arguments{
\item{res}{}

\item{perf}{}

\item{size}{}

\item{index.pnl}{}
}
\description{
Summarize the performance of serveral portfolio
}
\note{
TODO: need some work to standarlize it
}

