% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PriceAndDeltaUpdate}
\alias{PriceAndDeltaUpdate}
\alias{PriceAndDeltaUpdate.OptionPos}
\title{Price And Delta Update}
\usage{
PriceAndDeltaUpdate(object, ...)

\method{PriceAndDeltaUpdate}{OptionPos}(object, date)
}
\arguments{
\item{object}{An portolio object}

\item{date}{The actual as of date of the calculation}
}
\value{
Calcuate the closing price and delta based on black scholes model
}
\description{
Price And Delta Update
}

