% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PrintToHtmlTable.R
\name{PrintToHtmlTable}
\alias{PrintToHtmlTable}
\title{Print table to HTML table using the powerful xtable function}
\usage{
PrintToHtmlTable(table, include.rownames = F, cap = NA)
}
\arguments{
\item{table}{A data.frame object of something that can be convert to a table}

\item{include.rownames}{A boolean value to decide if the row names should be include in the output}
}
\value{
the return value is not important, this function is used for its side effect
}
\description{
Print table to HTML table using the powerful xtable function
}

