% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ThetaAdjNotional}
\alias{ThetaAdjNotional}
\alias{ThetaAdjNotional.OptionPos}
\title{Theta adjust notional value of a portfolio}
\usage{
ThetaAdjNotional(object, ...)

\method{ThetaAdjNotional}{OptionPos}(object)
}
\arguments{
\item{An}{OptionPos object}
}
\value{
the Theta Adjust Notional value
}
\description{
Theta adjust notional value of a portfolio

Theta Adjust notional value of an OptionPos object
}

