% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Add.R
\name{Add}
\alias{Add}
\alias{Add.EquityPtf}
\alias{Add.OptionPtf}
\title{S3 generic function  Add a object into a list}
\usage{
Add(object, ...)

\method{Add}{OptionPtf}(object, strike, expiration, type = c("C", "P"),
  underlyingName = "", name = "", multiplier = 100, amount = 0,
  openPrice = NA, price = NA, delta = NA, vega = NA, theta = NA,
  underlyingPrice = NA, underlyingSigma = NA, interestRate = NA, ...)

\method{Add}{EquityPtf}(object, name = "", amount = 0, openPrice = NA,
  price = NA, multiplier = 1, ...)
}
\arguments{
\item{object}{the object to be working on}

\item{...}{Other parameters}

\item{strike}{The strick of the new option}

\item{expiration}{The expiration date of the new option}

\item{type}{What is the type of the option: call or put}

\item{underlyingName}{The underlying ticker of the option}

\item{name}{The name of the security. For options, if missing automatically construct one based on the option information}

\item{multiplier}{The contract size of the contract. For equities, it is designed mainly for the purpose of contract, e.g. future}

\item{amount}{The quantity of the option contracts/equities}

\item{openPrice}{The opening price of the security (most probably yesterday's close)}

\item{price}{The current price of the security}

\item{delta}{The delta of the option}

\item{vega}{The vega of the option}

\item{theta}{The theta of the option}

\item{underlyingPrice}{The price of the underlying security}

\item{underlyingSigma}{The volatility of the underlying security}

\item{interestRate}{the current interestedRate}
}
\value{
The OptionPtf with the new option appended

The OptionPtf with the new option appended
}
\description{
S3 generic function  Add a object into a list
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPtf}: Append an new option object to the end of OptionPtf object

\item \code{EquityPtf}: Append an new equity object to the end of EquityPtf object
}}

