% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CumReturn.R
\name{CumReturn}
\alias{CumReturn}
\title{Cumulative Return, or the performance of a portfolio given the P&L}
\usage{
CumReturn(pnl)
}
\arguments{
\item{pnl}{The P&L}
}
\value{
The cumulative return
}
\description{
Cumulative Return, or the performance of a portfolio given the P&L
}
\references{
Performance2Pnl
}

