% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DeltaAdjNotional.R
\name{DeltaAdjNotional}
\alias{DeltaAdjNotional}
\alias{DeltaAdjNotional.EquityPtf}
\alias{DeltaAdjNotional.OptionPtf}
\title{Delta adjust notional value of a portfolio}
\usage{
DeltaAdjNotional(object, ...)

\method{DeltaAdjNotional}{OptionPtf}(object, ...)

\method{DeltaAdjNotional}{EquityPtf}(object, ...)
}
\arguments{
\item{object}{An portfolio object}

\item{...}{Other parameters}
}
\value{
the Delta Adjust Notional value

the Delta Adjust Notional value
}
\description{
Delta adjust notional value of a portfolio

Delta Adjust notional value of an OptionPtf object

Delta Adjust notional value of an OptionPtf object
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPtf}: 

\item \code{EquityPtf}: 
}}

