% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Drawdown}
\alias{Drawdown}
\title{The Drawdown of a time series}
\usage{
Drawdown(perf)
}
\arguments{
\item{perf}{An historical performance time series}
}
\value{
The draw down of the historical performance
}
\description{
The Drawdown of a time series
}

