% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MaxDrawdown}
\alias{MaxDrawdown}
\title{The maximium drawdown of a time seris}
\usage{
MaxDrawdown(perf)
}
\arguments{
\item{perf}{The historical performance of a time series}
}
\value{
A list includes \itemize{\item maxdrawdown: The MDD value \item from: The starting time of the MDD \item to: The end time of the MDD}
}
\description{
The maximium drawdown of a time seris
}

