% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MaxDrawdown.R
\name{MaxDrawdown}
\alias{MaxDrawdown}
\title{Maximum drawdown}
\usage{
MaxDrawdown(perf)
}
\arguments{
\item{perf}{The historical performance of a time series}
}
\value{
A list includes \itemize{\item maxdrawdown: The MDD value \item from: The starting time of the MDD \item to: The end time of the MDD}
}
\description{
The maximium drawdown of a time seris in the percentage values.
The maxdrawdown function in tseries
does not seem to be right given the fact that it is counting the absolute
value changes
}
\examples{
\dontrun{
require(quantmod)
google <- getSymbols("GOOGL", auto.assign = F)
perf <- Ad(google)
mdd <- MaxDrawdown(perf)
plot.zoo(perf)
AddMaxDrawdownArrow(perf, mdd)
}
}

