% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Notional.R
\name{Notional}
\alias{Notional}
\alias{Notional.OptionPtf}
\title{The Notional value of a portfolio}
\usage{
Notional(object, ...)

\method{Notional}{OptionPtf}(object, ...)
}
\arguments{
\item{object}{The portfolio object}

\item{...}{Other parameters}
}
\value{
The notioal value
}
\description{
The Notional value of a portfolio

The notional value of the given option portfolio object
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPtf}: Caluate the notional value of a Option portfolio
}}

