% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PriceAndDeltaUpdate.R
\name{PriceAndDeltaUpdate}
\alias{PriceAndDeltaUpdate}
\alias{PriceAndDeltaUpdate.OptionPtf}
\title{Price And Delta Update}
\usage{
PriceAndDeltaUpdate(object, ...)

\method{PriceAndDeltaUpdate}{OptionPtf}(object, date)
}
\arguments{
\item{object}{An portolio object}

\item{...}{Other parameters}

\item{date}{The actual as of date of the calculation}
}
\value{
Calcuate the closing price and delta based on black scholes model
}
\description{
Price And Delta Update
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPtf}: 
}}

