% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PrintToHtmlTable}
\alias{PrintToHtmlTable}
\title{import xtable}
\usage{
PrintToHtmlTable(table, include.rownames = F)
}
\arguments{
\item{table}{A data.frame object of something that can be convert to a table}

\item{include.rownames}{A boolean value to decide if the row names should be include in the output}
}
\value{
the return value is not important, this function is used for its side effect
}
\description{
import xtable
}

