\name{RBacktesting-package}
\alias{RBacktesting-package}
\alias{RBacktesting}
\docType{package}
\title{
A backtesting toolkit
}
\description{
Some tools to facilitate process of trading stratergies backtesting. It is not intented to replace any of existing sophistic backtesting packages like Rmetrics. It is just something that I need to use daily.
}
\details{
\tabular{ll}{
Package: \tab RBacktesting\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-10-17\cr
License: \tab What license is it under?\cr
}
This package includes functions belongs two the following section:
* Getting data
* Back testing data
* Evaluated portfolio performance
* Evaluate per trade performance
* Option portfolio
}
\author{
Bob Li

Maintainer: 
}
\references{
Packages depends on:
* quantmod
* ggplot2 
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
