% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SharpeRatio}
\alias{SharpeRatio}
\title{Calcuate the Sharpe Ratio}
\usage{
SharpeRatio(pnl, r = 0, freq = 1, na.rm = T)
}
\arguments{
\item{pnl}{The Sharpe ratio based on pnl}

\item{r}{The interest rate}

\item{freq}{The frequency of the sample. e.g. if daily data are used but to calucate annulized Shape ratio, freq = 1/365 or 1/265}

\item{na.rm}{A boolean value used to show if NA's should be removed during the calculation}
}
\value{
The Shape ratio
}
\description{
Calcuate the Sharpe Ratio
}

