% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ThetaAdjNotional.R
\name{ThetaAdjNotional}
\alias{ThetaAdjNotional}
\alias{ThetaAdjNotional.OptionPtf}
\title{Theta adjust notional value of a portfolio}
\usage{
ThetaAdjNotional(object, ...)

\method{ThetaAdjNotional}{OptionPtf}(object)
}
\arguments{
\item{object}{A portfolio object}

\item{...}{Other parameters}
}
\value{
the Theta Adjust Notional value
}
\description{
Theta adjust notional value of a portfolio

Theta Adjust notional value of an OptionPtf object
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPtf}: 
}}

