% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TotalValue}
\alias{TotalValue}
\alias{TotalValue.OptionPos}
\title{The total permium value of a portfolio}
\usage{
TotalValue(object, ...)

\method{TotalValue}{OptionPos}(object)
}
\arguments{
\item{object}{An portfolio object}

\item{object}{An OptionPos object}
}
\value{
The premium value of the portfolio
}
\description{
The total permium value of a portfolio
}

