% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TotalValue.R
\name{TotalValue}
\alias{TotalValue}
\alias{TotalValue.EquityPtf}
\alias{TotalValue.OptionPtf}
\title{The total permium value of a portfolio}
\usage{
TotalValue(object, ...)

\method{TotalValue}{OptionPtf}(object, ...)

\method{TotalValue}{EquityPtf}(object, ...)
}
\arguments{
\item{object}{An portfolio object}

\item{...}{Other parameters}
}
\value{
The premium value of the portfolio
}
\description{
The total permium value of a portfolio
}
\section{Methods (by class)}{
\itemize{
\item \code{OptionPtf}: 

\item \code{EquityPtf}: 
}}

